Fixed all the OSX Makefiles.
Added function to fetch image into ofxUtilities.H

Added exception classes to ofxUtilities.H

Modified all plugins that fetch images to check for failure, they then
report an error on fetch failure only if they are not being interrupted.

Fixed the basic gain example so that, when a general effect, if a mask is
not attached, it does no psuedo masking. Before it behaved as if it was
masked with black, which is a no-op. Which was silly.