Revert "Moved the parametric specific prop setting to ofxhParametricParam.cpp in the constructor of the ParametricDescriptor, as we need them at the time of the constructor to be able to set a hook on some properties."
This reverts commit 7386690c1ede148eb246d8d42c7961a218d53ca3.

The Descriptor class is generic and shall be used for all parameters.
Parameter-specific handling and data storage go to the instance.
The OpenFX doc mentions that the default control point should be stored
in the descriptor:
http://openfx.sourceforge.net/Documentation/1.3/ofxProgrammingReference.
html#id479272
« To set the default value of a parametric parameter to anything but
the identity, you use the control point setting functions in the new
suite to set up a curve on the descriptor returned by
OfxParameterSuiteV1::paramDefine. Any instances later created, will
have that curve as a default. »

This can easily be implemented using a private Double X 3N property on
the descriptor (let’s call it OfxParamPropControlPoints, and define it
in ofxhParametricParam.h, since it’s host-specific), where we store
(curveIndex, key, value) triplets. Only the control points need to be
stored. The descriptor does not need to handle the curve stuff : no
getValue, no tangents/derivatives, etc. It’s just a place to store
default values.

-> @MrKepzie