improve scripts/genRelease: allow test builds with dirty git.
scripts/genRelease is the script that builds OFX releases, builds and
packages the headers and doc, and uploads to the OFX Association
website.  When doing test builds it's inconvenient to have to have a
clean git working dir because you have to commit everything each time.
I added a --dirty arg to allow doing test builds.  Don't use this for
actual release builds of course.