fix casts between pointer-to-object and pointer-to-function
cast between pointer-to-object and pointer-to-function is only
conditionally-supported.
// Ref:
// ISO/IEC 14882:2011
// 5.2.10 Reinterpret cast [expr.reinterpret.cast]
// 8 Converting a function pointer to an object pointer type or vice
versa is conditionally-supported. The meaning of such a conversion is
implementation-defined, except that if an implementation supports
conversions in both directions, converting a prvalue of one type to the
other type and back, possibly with different cvqualification, shall
yield the original pointer value.