update to ofxParam.h -  ParamTypeClip
The purpose of this parameter is to logically locate in the effects controls the layer param indexer and is particularly useful for hosts that don't have other means to connect additional input clips to effects. The clip creation per se remains the same and the param name should match clip name.