ofxImageEffect.h
Adds better handling of spatial format

We add a new action to compute the format of the output clip, kOfxImageEffectActionGetImageFormat, which would ask the effect to set the appropriate properties on the outArgs.

    kOfxImageEffectImageFormatResolution a 2D integer indicating the resolution in pixels of the output image, this is always full res with no render scale applied,
    kOfxImageEffectImageFormatAspectRatio  ( Type - double X 1).  If not set assumed to be 1.0

For Generator we also add

kOfxImageEffectPropGeneratorType -  a string property that takes one of two values:

    kOfxImageEffectGeneratorFormatMaster
    kOfxImageEffectGeneratorFormatSlave

A generator of type kOfxImageEffectGeneratorFormatMaster will have the kOfxImageEffectActionGetImageFormat action called to determine the format of the clip being created, while a kOfxImageEffectGeneratorFormatSlave will not. A slave will have the format determined for it by the host and the relevant properties available on the output clip.

Generators that do not set this property will be defaulted to FormatSlave - the current behaviour (with spatial format inherited from  kOfxImageEffectPropProject* )
kOfxImageEffectActionGetImageFormat
 kOfxImageEffectPropSupportsImageFormat