OCIO Properties
A set of OpenColorIO-related ImageEffect clip properties which allow hosts to pass information about their OCIO configuration into a plug-in.

This allows plug-ins which display images in their own UI to match the colours used in the host, and is also useful in effects which are sensitive to the working colourspace of the input clip.

Signed-off-by: John-Paul Smith <jps@borisfx.com>