Reworked the OCIO extension to support ACES
The OCIO API has been extended to be a broader colour management API, although still using OCIO as the foundation.

This commit adds the ACESCG and ACES colour management styles, which only use OCIO as a source of colourspace names and could be implemented using other colour management pipelines.

OCIO-specific properties are separated and form a superset of the functionality offered by the ACES styles.

Signed-off-by: John-Paul Smith <jps@borisfx.com>