Update Example Makefile: fix Mac symbol visibility & add C++ std
Per @john-paulsmith's comments, symbol files are not needed when using
`-fvisibility=hidden` and the `EXPORT` macro. I deleted the symbol
files too.

We're using `--std=gnu++14` elsewhere, and some of the examples use
initializer syntax so this fixes the build.

Also added building Examples in the CI build so we'll know if they
break. Didn't do Windows yet, and the Support examples' Makefile also
needs some work, which is a separate issue.

Fixes #122.

Signed-off-by: Gary Oberbrunner <garyo@darkstarsystems.com>