Added the concept of core colourspaces and display names
The colourspace list has been updated to studio-config-v2.1.0_aces-v1.3_ocio-v2.3.

The Studio and CG configs are compared to add a new attribute “IsCore”. This is true if the colourspae is present in the CG config, and defines a smaller set of colourspaces.

Also, for most colourspaces, the first alias is a better short name than the actual colourspace name. The name is now used as the “DisplayName” attribute and the first alias is used as the base name instead.

genColour is now hard-coded to specific OCIO configs so it’s easier to do the Studio vs CG comparisons.

Signed-off-by: John-Paul Smith <jps@borisfx.com>