Added basic colourspaces instead of roles
Basic colour management mode no longer refers to roles and instead uses a small set of OFX-specific colourspace names which should cover most scenarios. These are defined in OCIO using roles for compatibility purposes, but in OFX, the expectation is that these might correspond to any colourspace that has the same encoding.

Signed-off-by: John-Paul Smith <jps@borisfx.com>