Support versioning native styles
Although the API should remain stable, colourspaces will change over time, and it’s likely hosts and plug-ins will support multiple versions of the spec for backwards compatibility. To address this, new properties have been introduced to allow negotiation of the config to use for native mode. This is somewhat similar to the use of built-in configs in OCIO.

Now that the native mode config negotiation is handled separately, the OCIO config property is reserved for use only in the OCIO style.

Hosts are now responsible for selecting the colour management style and config, and must set these in image effect properties.

Anticipating support for multiple versions, ofxColourspaceList.h has been renamed to a versioned name and is referred to as “the config header”. For now, it could be used as a drop-in replacement for ofxColourspaceList.h, but as soon as a second version is added, developers will need a strategy to manage these.

Additionally, the basic colourspaces have been renamed to include whether they are scene or display referred, and a descriptive label has been added.

Signed-off-by: John-Paul Smith <jps@borisfx.com>